Check for config file before fixing during before-save-hook
Only run php-cs-fixer-fix in the before-save-hook if our config file
exists.

If you set the config file option to 'project, then this will only run
php-cs-fixer on files that have a php-cs-fixer config file in the root
directory.

This let's me run php-cs-fixer on some projects but not others
automagically.